Refactoring router and service methods
