+++ server/b2b/router/__init__.py
@@ -1,1 +1,2 @@
-from .google_auth import auth_google_router
+from .auth_google_router import auth_google_router
+from .user_router import user_router
